- handle-errors added sqlx postgres to use the error list; added argon2 to use hash password error list; added custom errors for wrong password, when it's not possible to verify password, database queries, decrypt token, when unauthorized; - webapp added missing table creation; added new columns and account table; added account type; added authentication route; added authentication methods on store; changed answer and question routes to work with account sessions; added authorization to protect sensible routes;

+++ src/routes/question.rs
@@ -2,6 +2,7 @@
 
 use std::collections::HashMap;
 
+use handle_errors::Error;
 use tracing::{event, instrument, Level};
 use warp::hyper::StatusCode;
 
@@ -9,7 +10,7 @@ use crate::{
     store::Store,
     types::{
         pagination::{extract_pagination, Pagination},
-        question::{NewQuestion, Question},
+        question::{NewQuestion, Question}, account::Session,
     }, services::profanity::check_profanity,
 };
 
@@ -36,9 +37,12 @@ pub async fn get_questions(
 }
 
 pub async fn add_question(
+    session: Session,
     store: Store,
     new_question: NewQuestion,
 ) -> Result<impl warp::Reply, warp::Rejection> {
+
+    let account_id = session.account_id;
 
     let title = check_profanity(new_question.title);
     let content = check_profanity(new_question.content);
@@ -59,7 +63,7 @@ pub async fn add_question(
         tags: new_question.tags,
     };
 
-    match store.add_question(question).await {
+    match store.add_question(question, account_id).await {
         Ok(question) => Ok(warp::reply::json(&question)),
         Err(e) => Err(warp::reject::custom(e)),
     }
@@ -68,42 +72,58 @@ pub async fn add_question(
 // pay attention!!! the signature need to follow this order!!! param, store, item to be updated
 pub async fn update_question(
     id: i32,
+    session: Session,
     store: Store,
     question: Question,
 ) -> Result<impl warp::Reply, warp::Rejection> {
 
-    let title = check_profanity(question.title);
-    let content = check_profanity(question.content);
- 
-    let (title, content) = tokio::join!(title, content);
- 
-    if title.is_err() {
-        return Err(warp::reject::custom(title.unwrap_err()));
+    let account_id = session.account_id;
+
+    if store.is_question_owner(id, &account_id).await? {
+        let title = check_profanity(question.title);
+        let content = check_profanity(question.content);
+     
+        let (title, content) = tokio::join!(title, content);
+     
+        if title.is_err() {
+            return Err(warp::reject::custom(title.unwrap_err()));
+        }
+     
+        if content.is_err() {
+            return Err(warp::reject::custom(content.unwrap_err()));
+        }
+     
+        let question = Question {
+            id: question.id,
+            title: title.unwrap(),
+            content: content.unwrap(),
+            tags: question.tags,
+        };
+            
+        match store.update_question(question, id, account_id).await {
+            Ok(res) => Ok(warp::reply::json(&res)),
+            Err(e) => Err(warp::reject::custom(e)),
+        }
+    } else {
+        Err(warp::reject::custom(Error::Unauthorized))
     }
- 
-    if content.is_err() {
-        return Err(warp::reject::custom(content.unwrap_err()));
-    }
- 
-    let question = Question {
-        id: question.id,
-        title: title.unwrap(),
-        content: content.unwrap(),
-        tags: question.tags,
-    };
-        
-    match store.update_question(question, id).await {
-        Ok(res) => Ok(warp::reply::json(&res)),
-        Err(e) => Err(warp::reject::custom(e)),
-    }
+
 }
 
-pub async fn delete_question(id: i32, store: Store) -> Result<impl warp::Reply, warp::Rejection> {
-    match store.delete_question(id).await {
-        Ok(_) => Ok(warp::reply::with_status(
-            format!("Question {} deleted", id),
-            StatusCode::OK,
-        )),
-        Err(e) => Err(warp::reject::custom(e)),
+pub async fn delete_question(id: i32, session: Session, store: Store) -> Result<impl warp::Reply, warp::Rejection> {
+    
+    let account_id = session.account_id;
+
+    if store.is_question_owner(id, &account_id).await? {
+        match store.delete_question(id).await {
+            Ok(_) => Ok(warp::reply::with_status(
+                format!("Question {} deleted", id),
+                StatusCode::OK,
+            )),
+            Err(e) => Err(warp::reject::custom(e)),
+        }
+    } else {
+        Err(warp::reject::custom(Error::Unauthorized))
     }
+    
 }
